chore: Rename registry -> swap_encoder_registry
Took 14 minutes

Took 8 seconds
