+++ src/encoding/evm/swap_encoder/swap_encoder_registry.rs
